Fixes reverted commit
Running the process_string function from within another script would have resulted in an IndexError. This commit fixes the issue by adding a try-except block to catch any exceptions and ultimately do nothing about them.

The reason for this commit to be reverted was an overlook on indentation. This commit fixes that issue.